+++ core/services/ocr2/plugins/ocr2keeper/evm/abi_v21.go
@@ -13,13 +13,14 @@ type evmRegistryPackerV21 struct {
 	abi abi.ABI
 }
 
-func NewEvmRegistryPackerV2_1(abi abi.ABI) *evmRegistryPackerV21 {
+func NewEvmRegistryPackerV21(abi abi.ABI) *evmRegistryPackerV21 {
 	return &evmRegistryPackerV21{abi: abi}
 }
 
 // TODO: implement other methods as needed
 
 // UnpackLogTriggerConfig unpacks the log trigger config from the given raw data
+// TODO: tests
 func (rp *evmRegistryPackerV21) UnpackLogTriggerConfig(raw string) (i_keeper_registry_master_wrapper_2_1.KeeperRegistryBase21LogTriggerConfig, error) {
 	var cfg i_keeper_registry_master_wrapper_2_1.KeeperRegistryBase21LogTriggerConfig
 	b, err := hexutil.Decode(raw)
@@ -29,7 +30,7 @@ func (rp *evmRegistryPackerV21) UnpackLogTriggerConfig(raw string) (i_keeper_reg
 
 	out, err := rp.abi.Methods["getLogTriggerConfig"].Outputs.UnpackValues(b)
 	if err != nil {
-		return cfg, fmt.Errorf("%w: unpack getUpkeep return: %s", err, raw)
+		return cfg, fmt.Errorf("%w: unpack getLogTriggerConfig return: %s", err, raw)
 	}
 
 	converted, ok := abi.ConvertType(out[0], new(i_keeper_registry_master_wrapper_2_1.KeeperRegistryBase21LogTriggerConfig)).(*i_keeper_registry_master_wrapper_2_1.KeeperRegistryBase21LogTriggerConfig)
